Latest version
I rewatched the lectures and realized that
I was probably doing the algorithm wrong, and in
a way that was much more complicated than we were
expected to do. I still need to write the code that
will actually export the features to the file, but
the SAOLA_test2.py script should produce a feature set.